Put localconfig VTS test output in /tmp
Instead of a file in the repo. Also remove the file when exiting script.

Would be better if we had some other way to check for success/failure (e.g.
by checking exit code from vts-tradefed) but investigating that is
something we can do later. At least we do not leave any stray files around
after this change.

Change-Id: Id639f54dacdf29db897e1f9f642fd2a3f2b7af2f