Add .clang-format
Seems to be a rough consensus on that we should follow Google's coding
guidelines. See https://google.github.io/styleguide/cppguide.html .

Our screens are wider than the screens of mainframes in the 1960s though
so increase column limit to 120. Let us change it later if there are any
objection to this. Just want to have a .clang-format in the repo.

Change-Id: Ic51242ecfba60dafb60c64bb954fe8ccdf82504d