Make sure values are of correct type in localconfig
JsonCpp in Android is built with JSON_USE_EXCEPTION set to 0. This causes
JsonCpp to call std::exit() if it cannot convert a value to the type
requested by caller of Json::Value::as*(). We do NOT want this behavior
in a library.

If JSON_USE_EXCEPTION is set to 1 JsonCpp will instead throw a
std::runtime_error(). But since we want to be a bit more strict (should
e.g. not be possible to use a bool as an int, which JsonCpp::asInt()
allows), check type explicitly and throw std::runtime_error ourselves if
value is of wrong type.

Change-Id: Icc22230734a480cb4d0ca6859710081d7a45ec23