Update docker_build image ref
For changes in d622b928d1a0f728b72cbc54ad028958a35dadd4 .

In the future we should perhaps use something other than the commit hash
as an id so that image.ref can be updated in the same commit as the
dependent changes are made in. Could use the change-id or just a randomly
generated number. (If image.ref is updated in the same commit, we have a
connection between the id and the commit.)

Change-Id: I1ff7eeb0b8fbfa1845ac050062cf6d188db4b66a