Pass all env vars set in Jenkins to Docker container
tools/docker_build/run.sh is modified to filter out --env-file from its
arguments.

Can only handle a single --env-file argument at the moment. The current
filtering also breaks if any of the arguments actually need to be passed
through to the command that should be executed in the container. Was like
that before as well though. If we need to fix any of these deficiencies we
should migrate to getopt or something.

Change-Id: Iabdfa398eca55c889fbd90532995b9edd1f6a129