Replace vendorsetup.sh with tools/envsetup.sh
Since it is not possible to get vendersetup.sh sourced through
build/envsetup.sh when we do not have a full source checkout. This is the
case when running on slaves connected to rigs.

Also only source build/envsetup.sh and tools/envsetup.sh in docker image
if they exist to allow for docker image to be used without any source code
checkout at all (e.g. when invoking it in Jenkins jobs before anything has
been cloned at all).

Change-Id: If943907d331a19834bdfea658f72144a0e503a08