Move sync:ing of repos required for VTS in gate jobs from Jenkins to repo
If what we need ever changes, we want it version controlled.

There is a lot of docker_run going on in the Jenkins scripts now. Ignore
it. Will be fixed at some later point when we move the bulk of what is
done in tools/ci/jenkins/*.sh to Python.

Change-Id: Ib610bab51c3383af38b707c8227b80fb0080481c