Fail when updating manifest repo if no changes are found
Had problems with zuul-cloner falling back to master. Tests in gate test
did not test what they should test and auto bumper did not auto bump.

Should be no case in gate build or auto bump jobs where we do not have
changes in the manifest repo. Make sure we fail if that is the case.

Change-Id: Iaddfd5a3fd6865c0548a5287156c8193e58c381f