Include unit tests mk-file from parent mk file
This is required to be able to run "mma" from the unit test directory,
otherwise MODULES-IN-**** will be undefined and the build system will
not find the module.
The aops main Android.mk file will only include makefiles recursively that
does not have any Android.mk in any parent directory, that's why the
components main Android.mk-file is included but the test is not.
(Strangely, "mm" works without this fix.)

This is part of a larger effort to standardize the way tests
are built and run. All tests must be able to build with mma.

Change-Id: Ie9e0e5c806d6bec8e45cb7473d7d22df3c805e38