Fdxclient - Lock Item while reading and writing
For receive-signals, write is done by fdx-thread
and read is done by main test-thread, hence the lock is required.

Also add close-function to FRSignalInterface.
If you don't close the receiver it will leave a danglig thread.
Even though the thread is marked with daemon=True, VTS scans for
running threads when the test is done and then takes very long to complete
and prints tons of warning messages.

Change-Id: I9f319cd00f466059d94e1485819a979af2a50560