Added netboy daemon which moves nics to vcc namespace.
The netboy daemon is run in the default namespace and listens for
RTM_NEWLINK events. When a new network interface has been detected
it is moved to the vcc namespace. Once the network interface has
been moved to the vcc namespace, netboy then notifies netman of the
new network interface by doing an ifconfig up. This is the only way
(that we know so far) to notify netman that a new interface has
been moved to the vcc namespace. Netman then receives the notification
and configures the network interface in the vcc namespace.

Netman runs in the vcc namespace as all of the configuration needs
to be done in the vcc namespace. Otherwise we would have to wrap
the configuration of network settings to make sure the configuration
was executed in the correct namespace.

By having netboyd only do one task in the default namespace also
ensures lower response times when moving interfaces to the vcc
namespace.

Change-Id: Ie32cd1bb47895ca4986eae90b2a1c0c79d9085d6