Fixed ip configuration tests to work with namespaces.
Most commands in netman_helper.py now run in vcc namespace by default.

Removed meth0 interface tests as the most driver is not operational.

Tidied up parts of the code.

At the moment we are using sleep methods in parts of the tests, as it
takes a while to fire up netman and to bring up/down interfaces. This
should change to a more robust way deterministic way.

Change-Id: I838c038d45a75a40810286c24160f254ab2c544f