Pass all Android build system and ccache env vars to Docker with --env
Instead of through --env-file in Jenkins builds. --env-file sets the
variable to the empty string if it is not set in the host env. --env leaves
it unset in the container if it is not set in the host env. Causes trouble
with e.g. OUT_DIR (builds break if it is "").

Change-Id: I76a83f7e0aa30385d16a92d299bd837627b327e4