First commit of an empty VCC VHAL daemon.
PSS370-3924
This is using the Delphi generic Vehicle HAL implementation and
the Base Module system.Each modules implements a number of
properties and in the VehicleHALServiceMain.cpp we decide
which modules to instantiate.
See https://c1.confluence.cm.volvocars.biz/display/IHUA/Vehicle+HAL
for information about the design.Manifest override is
to include latest VCC change in Delphi repo which this change
is dependant on.

Change-Id: I04f1fa9f835c0b55bec3e137daa7527ee0ab23d1