Check for Python test cases
There was no way to see if the tests hade run or not,
this will copy and parse test_run_summary.json to check
python results.
It will only check the results if there is a json file
and the test module running matches the json file. And
will only generate exception if there is fail or a test
case that is not run. There are other test cases (gtest,Java)
that do not generate at json file with test summary.

Change-Id: Ic24f6097fb4cfa386342b4dcf123a543e07e1e28