Use ZUUL_COMMIT when pushing/pulling build result to/from Artifactory
ZUUL_CHANGE_IDS can apparently change between a build and test job.
Suspect it happens when a speculative merge is done and the base commit is
merged before testing starts for the speculative one.

ZUUL_COMMIT is the Git hash of the commit that is being built and tested.
Should not change for a particular build/test run regardless of what has
happened with previous commits.

Change-Id: I64a1e33ebd29ea6e7792dfd5f42d744e2fd0c0d8