Move adding of VTS test case path to PYTHONPATH from shell to Python
So it is set correctly when invoking test from Python as well.

Forward kwargs down to asyncio.create_subprocess_exec() so "env" can be
used to set a custom environment for a command.

Also copy smoke test to gate (seems stable enough and actually tests that
this commit works) and move it to top of hourly to... smoke test... before
running any other tests.

Change-Id: I623e8f3e22e60dfdbd71f0266463b164a4888e8f