Fix possible deadlock in hidl notification/getService
The implementation of onRegistration and getService seems to have changed
in O.MR1. In the new implementation getService() will try to get the
service once and if it isn't found it will register a notification handler
and wait in a loop until service is available. Dataelements also registers
a notification handler but before the call to getService() resulting in
that callback handler being called before the handler registered by
getService(). The mutex in connectToVsdProxyAndResend will then prevent
the callback method from continuing, resulting in a deadlock in dataelements.

By removing the call to connectToVsdProxyAndResend, this is avoided.

Change-Id: I16882e2479c3af25ddabb208aee5bd64b48bbde4