Added ip command to vendor partition.
A simple copy operation from /system/bin to /vendor/bin is
done during build.

The advantage of using Android's ip command is that we get
the (possible) security updates from Google. And that we
don't risk breaking anything in Android, as we're using
standard commands for our functionality.

The downside is that we rely on Google that they'll support
namespaces in the future.

Change-Id: I9d0e60fb0c4bef535ddae2678d409108070bbbac