Use Car Keyboard instead of default
device.mk is updated to build the CarLatinIME apk which
is the specific Car keyboard.

There are resource overlays added in order to optimize the
keyboard layout (distances between keys) to fit the IHU target
resolution and screen size.
Both the regular and the symbol keyboards are updated.

This will build the CarLatinIME apk so that
the Car Keyboard can be used instead of the default.

Change-Id: I021fb77c8c53b60e605d134a31cb9e79d8e489d5