Configurable skip of VNDK api checks during build
This (ugly hack) speeds up the build almost x2 on servers, which is the only
reason for its existence. The speedup comes at the price of sacrificing the
VNDK ABI checks probably added by google in MR1.
The "ugly hack" part is that we're actually modifying the soong scripts
directly. The reason for modifying run.sh in this way is to give developers
a way to use the speedup that this hack provides without it ever making its
way into an upstream patch. Hopefully this will be a shortlived workaround
and we will either find a better way to handle this or get google to make
this configurable somehow.
There is no story reference for this at this time but it is part of
epic [PSS370-12878].

Change-Id: I1e6d1043616656a759be30175c0c0acb7e3edee8