Add AOSP build root as a docker volume
Currently if the docker run script is executed in a subdirectory of the
AOSP build root it will only include your home directory and your
current working directory as volumes for docker. This means that if the
directory that contains the AOSP build root is not located in your home
directory, the docker container will not be able to use e.g the binaries
in the prebuilt directory located in the AOSP build root.

To resolve this, the run script has been updated to always include the
AOSP build root as a volume.

Change-Id: Ie087f1ea5a26d47198a98595b3d07709d1d543bd