Add year and old file handling to copyright script
Currently the copyright script is hard coded to only handle the year
2017 and if any other year is written in the header it is deemed invalid
and the header is overwritten with a new header containing 2017.

To resolve this, the script has been changed to fetch the current year
instead of using the hard coded value. It will also take the old
copyright year in headers into account when updating Volvo files, so
instead of just overwriting older headers with a new one containing the
current year. It will write a header with a range from the original
copyright year to the current one.

Change-Id: Ia5fa51fd09f0b358984bdef04fdbbcea7ffc9c8f