Fix docker handling of non matching ttyUSB glob
If there is no matching files when using a glob pattern, the pattern
itself will be the result. As the pattern is not a valid file the result
needs to be checked before it can be used. Currently the entrypoint
script for docker does not perform this check when it globs for ttyUSB
devicesi. Thus it tries to change the group ownership of the glob pattern
when no match is found, which results in an error.

This has been resolved, by checking that the result is an existing file
before trying to change its group ownership.

Change-Id: Ia31b0ad321c262440fe48669e2b332673caf0207