Adds docker system prune to both clients and CI machines
Docker provides a command that ensure that dangling
images are removed and one could perform a system
clean-up everynow and then. This commit provides
an interface for both a developer and CI machines
running a build container.

For CI machines the system is cleaned up everytime a
container is started.

For interactive users/developers the feature is disabled
by default and can be enabled if needed.

[ARTINFO-818]

Change-Id: I08554baddc9d449a4cd5aed0b3c0376b712290a1