Add HMI commit check
Add a commit check for HMI repos that support Gradle.
The commit check is intended to provide quick feedback by
building, running unit tests on host and doing static code analysis.
If any of the quality gates fail, Jenkins should set a -1 Verified
status.

Tests and analysis are run through Gradle's check task, and it's
up to each project to provide tests, which type of analysis
should be run and how the output should be treated.

[ARTINFO-2518]

Change-Id: I5007a75065edaa27d76076d7cd4f1dc5328b64ee