Update statusbar height incl. space for top bar
Increase the height of the statusbar to include both the
height of the statusbar plus additional space for the top
navigation bar (according to latest design for Volvo). The
extra spacing in the status bar layout should be removed
when the top navigation bar can affect the size of the
application window on its own.

[ARTINFO-1052]

Change-Id: Ie79c29885623e613f277cc3a91d68fb477b3aa0c