Added "vsomeip" directory creation under /dev/vendor
VSOME/IP software stack uses unix sockets for communication between
stubs and proxies. This is inline with Android's (SEAndoid) expectation
where inet socket communication and binder communication is not allowed
from same daemon.

Socket fds for UNIX sockets needs to be placed on a tempfs and so are
placed under /dev/vendor/vsomeip

Change-Id: I2771aa2145026276152fda7677358903aaea9e75