Change VtsPythonVirtualenvPreparer to multi_target_preparer
In Android P, VTS started throwing errors when using target_preparer
for VtsPythonVirtualenvPreparer.
All of the VTS examples using VtsPythonVirtualenvPreparer in AOSP were
changed from target_preparer to multi_target_preparer at the same time.

This is because VtsPythonVirtualenvPreparer is preparing the host and
not the target, multi_target_preparer reflects that better.

Change-Id: I4516f9f369493c36e69bee2d3f463b351d2ed1db