Don't wait for IHU restart after starting canoe in test cases
Older versions of the VIP would restart the whole IHU when
connecting/disconnecting the flexray interface.
That no longer happens.

Also, the stopServices/waitForDevice/startServices is quite
unreliable.
stopServices() will internally setprop sys.boot_completed to 0
and then we call waitForBootCompletion which waits for sys.boot_completed==1.
So if the flexray reconnect doesn't actually restart the device it will
be stuck with sys.boot_completed=0 and waitForBootCompletion will block
for the entire default timeout which is 15minutes.

Change-Id: I78f9c4796c60aaa45cf3331549034633d61ef590