Fix copyright regex for python's encoding comment
The regex in the copyright script that detects Python's magic "coding"
comment that specifies the Python file's encoding. Only matches the
magic comment when it uses an equal sign, however the syntax allows it
to be written with a colon as well. This caused the copyright script to
not detect the valid copyright headers in files using the magic comment
with a colon and an additional copyright header was added.

The new regex taken from PEP 263, matches all the valid formats of the
magic "coding" comment.

Change-Id: I6996eac2ef95db8a3078e62c9ffa81dff4212fed