Fix regex for copyright without trailing newline
The current regex for hash comment styled copyright headers assume that
there is a trailing newline after the copyright header. This causes the
script to not detect the copyright headers if there is no trailing
newline and thus it adds a new copyright header above it.

The regex has been changed to detect the copyright header even without
the trailing newline and the existing code to fix copyright headers will
detect the missing newline and add one.

Change-Id: Ifff62c6e300aec09b4879d797f7f487d7d890d89