Add checks for non Volvo and multiple copyrights
Currently when the copyright script fixes copyrights in files it doesn't
check if the file already contains a non Volvo copyright header and
instead just adds a new Volvo copyright if one is not found. This has
caused the script to add a Volvo copyright header to files that already
have a different copyright header.

Files that have a non Volvo copyright shall be placed in the external or
thirdparty directory, the script has therefore been changed to show an
error if it finds a file with a non Volvo or multiple copyright headers
out side those directories.

Temporary exceptions have been added for the paths that currently
violates the rule about where to place non Volvo copyrighted files and
JIRA issues has been created for the respective teams.

Change-Id: I8772ee6b3ff4244b38ec2ea20ccb2e8e572ae8d0