Remove unsubscription from audiomanager in soundwrapper dtor
This is to prevent a crash originating from creating a new temporary sp
pointing to the already to be destroyed instance, when this sp goes out of scope
it triggers the destructor again creating a stack overflow and crashes.

Unsubscribing "this" from am_service should not be needed in
SoundWrapper destructor, since the only way to reach dtor is that
all sp pointing back to "this" are already dead and hence there
is no valid pointer left to the SoundWrapper instance that can be
used to trigger a callback.

[ARTINFO-2768]

Change-Id: Ie658d665cb64b6f1401f7ea08dea6567df78f9e4