Remove drive mode fault handling from climate defroster functions
At the moment IHU does not distribute DriveMode flexray signal. This causes the front and rear defroster to not work in a real car.
Temporary remove the fault handling of DriveMode signal to enable this function. This "hack" will be removed in Jira issue ARTINFO-2827.

Partly resolves ARTINFO-2094 - Rear defroster, Electric Defroster - HMI

Change-Id: Ie45b283fcb089f458e7d6ebf16559b68e5b77c99