Remove Vector config from env.list
It is now being read from lab_config.json instead.

Having the variables in the env.list causes issues because when
an environment variable is unset on the host it will become set to
empty string inside the docker. We have some scripts that make distinction
between undefined and empty env variable.

Change-Id: Ia6c89c929d7e748f09584c9364b0b076345329f5