Cleaning workspace if repo init/sync fails
repo init and repo sync sometimes fails if the workspace isn't clean.
That can happen if there are uncommitted files in the wrong places.
Instead of cleaning the workspace in every build which works we only
need to do it when it fails.

On normal builds this will save ~10 min. On failing ones this can in
worst case cost ~10 min. Hopefully the good cases are the vast majority.

Change-Id: I7b2bfbbf9ca8a13660d69229c8ac09d530efb657