Integration Test for Sound Notifications
Initial commit for the IT of sound notifications,
it includes test case for park brake sound,
Turn indicator and Belt reminder sounds but right now
its pushed just to check if its executing on the
CI environment.

It has the dependency on audio alsa recorder, audio
matching algo and audio command line utilties.

Audio record and matching scripts are present on the Gerrit:
https://icup_android.gerrit.cm.volvocars.biz/#/c/5093/

part of work ARTINFO-2326

Change-Id: I63c3caf316eb37a638eb23bba5d20aae73467b6b