Add base for camera resource Mgmt to EVS manager
Adds bare implementation of the IEvsCameraStream and IEvsCamera
HIDL interfaces to the EVS manager. The IEvsCameraStream receives
the video stream from the hardware camera implementation and
distributes it to one or more virtual cameras.
The virtual cameras implement the same IEvsCamera interface as the
hardware camera and acts as a proxy to the client.
The end goal is to allow video from a single physical camera to be
shared among one or more clients, while allowing the clients to act
as if they had direct access to the physical camera.

Tracked-On: ARTINFO-2885

Change-Id: Id87523775d718aacbb4d6e3f6be6c2be973ed11d