PowerModule answers with the wrong alive reason
The VIP can request the current alive reason from the MP.
This request is received by the PowerModule and is now
always answered with the incorrect, hardcoded, value
of zero (0). It should instead ask the LifecycleControl
instance to send the response.

Argument parsing fix in iplmd

This fix shall also update the volvo manifest and needs:
https://icup_android.gerrit.cm.volvocars.biz/#/c/5685/

Change-Id: Ieb76c6945355abbb7d84209fca85e1ce324c4f24