Set success in conditional step
Our multijobs are set to continue to next step Always. That means that
a step after will run even if the job has failed. Setting of status needs
to be set only on success and we therefore need to use a conditional step.

Also moved some stuff for testjobs to the default.

Change-Id: I7e0876840b9a86741c41fc6fffe52b5a604be0bb