Add compiler message format validation in libdbg
To make it possible for the compiler to validate the message format
passed to the log function in libdbg. The compiler needs to know that
the log function uses the printf format style. This will detect
incorrect format types and possible security problem from passing
something that is not a string literal as the message format.

Change-Id: I120b1440c5ba9cc363889afa7c6292e8b6da164b