Add remote climate service
The service receives messages from the gateway and further
communicates with the VHAL to get/set properties. It then
sends the response back to the gateway.

The service is a part of the convenience API used by third
party applications to remotely control the climate in the
car.

Part of work on ARTINFO-2243

Change-Id: Ie281d007c45919fd8d1c21b8f9a251a7f05481fb