Always black status bar
Sets the status bar to always be black, regardless of what
color applications try to color it with.
Note that this may not play well with apps that want to
draw behind the status bar, since the apps still will take
space behind it, but not being able display their color
and/or content - this will have to be fixed later on.

[NO JIRA]

Change-Id: I353a0adddb1335df257f3afcf03432be033537a3