Refactor tester.py
As the script tester.py has grown to handle lots of stuff we needed to refactor it somehow.

Important note is that this cannot be merged as is as there has been many conflicting changes that are not included. If this looks good we'll do a new refactor with the current code.

Change-Id: I3f7edcb3f517ccb4c8a8b7a8ad2619678c503296