Update notification overlays
Set the notification text color to match the dark style. Note
that the text color was resolved automatically in O MR1, so
once P stabilizes we may not need the text color overlay.

Use the Volvo font for notification texts. We overlay the
Material Style temporarily to make this happen. This will
not pass CTS, and long term we should change so that notifications
use the DeviceDefault theme instead.

Patch depends on Ib196002c3ba0c209bb95c2c3d3973d7ecddf87f1

Change-Id: I5ea01cb19594605b98aa975bcae8dbf782323d88