Add interface for EVS manager virtual cameras
Adds an interface, IVirtualCamera, that extends the IEvsCamera
interface. The interface will allow for easy mocking of VirtualCamera
objects, which will in turn allow for good unit test coverage.

The virtual cameras implement the same IEvsCamera interface as the
hardware camera and acts as a proxy to the client. The end goal is to
allow video from a single physical camera to be shared among one or more
clients, while allowing the clients to act as if they had direct access
to the physical camera.

Change-Id: I72f7dfeecd9c47fffca04a1d3571296619fff3bf
Tracked-On: ARTINFO-2885