PAC: Rename implementation of IEvsCameraStream
Renames the wrapper interface for IEvsCameraStream,
IEvsCameraStreamWrapper, to IEvsVideoProvider. Also renames the
associated implementation, EvsCameraStream, to EvsVideoProvider.

This will hopefully make it clearer what the class and its
associated interface are responsible for.

IEvsVideoProvider implements the IEvsCameraStream HIDL interface. The
IEvsCameraStream receives video from the hardware camera implementation
and distributes it to one or more virtual cameras. The virtual cameras
implement the same IEvsCamera interface as the hardware camera and acts
as a proxy for the client. The end goal is to allow video from a single
physical camera to be shared among one or more clients, while allowing
the clients to act as if they had direct access to the physical camera.

Change-Id: I30729098e4b2cf2bca7b2aea16ece2627dff9456
Tracked-On: ARTINFO-2885