Show failure reason in build description for icup_android_manifest_bump
Use description-setter plugin to add the reason for failure, by matching
"fatal:" in the log output.

To make this work, some cleanup in log output is needed, since "fatal:"
also could be written due to a redundant command.

When a repo already exists, and origin is defined, an error message
appears in the log which is always to be ignored - e.g:

 + git init
 Reinitialized existing Git repository in ...
 + git remote add origin ssh://gotsvl1415.got.volvocars.net:29421/manifest
 fatal: remote origin already exists.
 + true
 + git fetch --tags ...

This change aims to avoid git remote add if not needed, and thereby make
use of the description-setter to show relevant error info as build description
by matching "fatal: "

E.g. (https://icup_android.jenkins.cm.volvocars.biz/job/icup_android_manifest_bump/3260/console)

 + git fetch --tags --progress origin dev/hmi_preP_ww11
 fatal: Couldn't find remote ref dev/hmi_preP_ww11
 Build step 'Execute shell' marked build as failure

Change-Id: I8996c019aca45bbcb6f69290fcb9d6a346e4825e