Include screenshots in test report
Also major restructure of how large blobs of test results,
such as screenshot and log-files are stored.
Instead of storing in the same document as test results they are stored
in a separate document with only bare minumum metadata. The test result
documents only contain an ObjectId pointing to the large log.

Change-Id: Ia98c53a058eed71376ad70f821661cdb66f114c8