Run SettingsStorage in single thread only
ProfileManager callbacks are now put on queue and replying OK instantly.
This avoids potential deadlocks when calling hidl-calls inside other hidl-calls
as this can create a circular or recursive call-chain between two processes where
the second process is not able to reply as it is busy waiting for an answer on
the first hidl call.

Resolves ARTINFO-3349

Change-Id: I4529a70c71c0e0bde93b7418a2b0953d0c4973de