Add video stream start/stop support to EVS manager
The EVS manager shall handle the distribution of video from the hardware
camera to one or more clients, using a virtual camera implementation so
that the clients can act as if they had direct access to the hardware
camera. To facilitate this it must be possible to start video streams
when they are requested by the clients and to stop them when they are no
longer needed.

This patch will allow clients to request a virtual camera to start and
stop a video stream.

Note that no actual video will be provided until frame handling support
is added. This will be done in a later commit.

Virtual cameras will forward start/stop video stream requests from
clients to the video provider. The video provider will forward video
from the hardware camera to the client via the virtual camera. The
video provider will ensure that the hardware camera stream is only
running when in use by a client.

Change-Id: I5cc03fea31bc8af5c5206f9d483c4ce6219812dc
Tracked-On: ARTINFO-2885