Rename static methods in dispatcher to prevent misuse
C++ has the "nice feature" that if you call a static method
on an instance of an object it will silently call the static method
of the class. The static "convenience" functions of dispatcher,
with almost exactly the same name as the instance methods, will then
affect global state instead of just the dispatcher instance you thought
you were operating with.

Rename the static methods to avoid such confusion between Enqueue vs EnqueueTask

Change-Id: I1b2463b9745bae8f9d926be89783f98dc2df5082