Add EVS hardware service registration retry loop
Because the EVS hardware service is started so early in the boot
sequence, the Android service manager may not be ready when the EVS
hardware service tries to register itself. This will cause the
registration to fail and the EVS hardware service to stop.

To resolve this, a retry loop has been added to allow the EVS hardware
service to handle this situation gracefully and wait for the Android
service manager to get ready.

Tracked-On: ARTINFO-3459
Change-Id: I7d02ae70031c3c7d81357ee0f5110eed9acba345