Add EVS manager service getservice retry loops
Because the EVS manager service is started so early in the boot
sequence, the Android service manager may not be ready when the EVS
manager service tries to retrieve the EVS hardware service. This will
cause the retrieval to fail and the EVS manager service to stop.

To resolve this, a retry loop has been added to allow the manager
service to handle this situation gracefully and wait for the Android
service manager to get ready.

Tracked-On: ARTINFO-3459
Change-Id: I00910731163d58b9804964e54b2ed51fabaeccc5