Add PAC early app getservice retry loop
Because the PAC early app is started so early in the boot sequence, the
Android service manager may not be ready when the early app tries to
retrieve the EVS manager service. This will cause the retrieval to fail
and the early app to stop.

To resolve this, a retry loop has been added to allow the early app to
handle this situation gracefully and wait for the Android service
manager to get ready.

Tracked-On: ARTINFO-3453
Tracked-On: ARTINFO-3459
Change-Id: I7f08f48db73c8a562c8aba71b535c0852bc86ae5