Add basic frame handling to EVS manager
The EVS manager video streams would be more useful if they actually
provided video.

This patch will allow video from the hardware camera to be forwarded to
the client requesting it.

Documentation is still poor and will be improved in a later commit.
Handling of edge and error cases will be improved in a later commit.

Frames from the hardware camera are now forwarded by the
EvsVideoProvider to all virtual cameras that are currently streaming.
The virtual cameras buffer the incoming frames and forward them to the
clients.

When a client returns the frame, the virtual camera removes it from the
buffer and returns it to the EvsVideoProvider. The EvsVideoProvider
returns the frame to the hardware camera when it is no longer in use.

Change-Id: I201b44a2ef0f8e5d6ab0ab8e7e61ad12f70af4b4
Tracked-On: ARTINFO-2885