Change EVS HW service display management behavior
Currently the EVS hardware service will kill an active display when
someone requests to open a new display as its implemented in Google's
reference implementation. However, the documentation for the EVS
Enumerator HIDL interface, states that if the display is already in use
or otherwise unavailable the request shall return a null pointer.

This commit changes the EVS hardware service's behavior to comply with
the documentation rather than the reference implementation.

Change-Id: I5a76b6fe85c638e50887673d9d31f4f9fcfddb14