PAC: Add doxygen documentation to EVS manager
The EVS manager video stream functionality is currently poorly
documented. Poor documentation complicates maintenance and adding new
functionality since understanding the exisitng code is a prerequisite
for changing it.

The video stream functionality in EVS manager is now documented using
doxygen-compliant comments. This should reduce the time required to
understand the intention of the code significantly.

Tracked-On: ARTINFO-2885

Change-Id: I8b5a3a3f192b2b8dc21670cf6ed56ec878fb9cc3