CI: Bugfix for dryrun-hourly
The build-script gate_build.sh was used previously, invoked by the
Jenkins job ihu_dry_run_build-eng. This was not sufficient - e.g.
some tests were not built.

With this change, the Jenkins-job may instead call the ordinary script
for hourly builds, hourly_build_eng.sh, with "dryrun" as a parameter
(which is needed for some specific actions for open changes)

Likewise, the ordinary script for hourly-test is used, called with the
parameter "dryrun", e.g. to control fetching from artinfcm instead of
artifactory. Thus, the script dryrun_hourly_test.sh is not needed.

Change-Id: Icaae1f359e7590931918885f1817f43318cf855f